legacy_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0
Flask==3.0.0
flask-cors==4.0.0
openpyxl==3.1.2
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.0
deep-translator==1.11.4

# Security Libraries
bcrypt==4.1.2
cryptography==42.0.2
slowapi==0.1.9
python-magic-bin==0.4.14
PyJWT==2.8.0
email-validator==2.1.0

# Google API libraries for Google Meet & Gmail integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Invoice & Bank Reconciliation Libraries
pandas==2.1.4
reportlab==4.0.7
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# System Monitoring
psutil==5.9.8
